saving

+++ app/forms/item_form.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SelectField, IntegerField
+from wtforms import StringField, SelectField, IntegerField, FloatField
 from wtforms.validators import DataRequired, ValidationError
 
 
@@ -7,7 +7,7 @@ class ItemForm(FlaskForm):
 
     name = StringField('Title', validators=[DataRequired()])
     # user_id = IntegerField('')
-    price = IntegerField('Price', validators=[DataRequired()])
+    price = FloatField('Price', validators=[DataRequired()])
     category_id = SelectField("Category", coerce=int, validators=[DataRequired()])
     description = StringField('Description', validators=[DataRequired()])
